storageminer: separate sectorstore types
